requirements.txt
asgiref==3.7.2
Django==5.0.2
mypy==1.8.0
mypy-extensions==1.0.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
psycopg[binary]==3.1.18
pytest-django==4.8.0
